基础架构02

+++ qk-dam/qike-dm-metadata/src/main/java/com/qk/dm/metadata/repositories/MtdLabelsRepository.java
@@ -3,6 +3,6 @@ package com.qk.dm.metadata.repositories;
 import com.qk.dam.jpa.base.BaseRepository;
 import com.qk.dm.metadata.entity.MtdLabels;
 
-public interface MtdLabelsRepository extends BaseRepository<MtdLabels, Integer> {
+public interface MtdLabelsRepository extends BaseRepository<MtdLabels, Long>{
 
 }
